Ok ws and db

+++ src/main/java/eu/goldenk/entities/Firmware.java
@@ -1,14 +1,13 @@
 package eu.goldenk.entities;
 
 import javax.persistence.*;
-import java.util.Date;
 
 @Entity
 public class Firmware {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private int id;
+    private long id;
 
     @Column(name ="name", nullable = false)
     private String name;
@@ -17,11 +16,11 @@ public class Firmware {
     private int version;
 
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
